pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["sanswed <sanswed@gmail.com>"]
readme = "README.md"
packages = [{include = "task_manager"}]

classifiers = [
    "Programming Language :: Python :: 3.8",
    "Framework :: Django :: 4.1",
    "Framework :: Flake8",
    "Topic :: Education",
    "Topic :: Internet :: WWW/HTTP :: WSGI :: Application"
]

[tool.poetry.dependencies]
python = ">=3.8.0"
gunicorn = "^20.1.0"
django-bootstrap4 = "^23.1"
django-filter = "^23.2"
rollbar = "^0.16.3"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
psycopg2-binary = "^2.9.6"
parameterized = "^0.9.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
coverage = "^7.3.0"
dj-database-url = "^2.0.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
